infra/karpenter-nodepool-cpu/images/onnx-embedder-reranker/pyproject.toml
[tool.poetry]
name = "rag8s-onnx-embedder-reranker"
version = "0.0.1"
description = "RAG8s ONNX embedder + reranker"
authors = ["platform <platform@example.com>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
ray = {version = "2.48.0", extras = ["serve"]}
onnxruntime = "1.22.1"
transformers = "4.55.2"
tokenizers = "0.21.4"
grpcio = "1.74.0"
grpcio-tools = "1.74.0"
numpy = "1.26.4"
uvicorn = "0.20.0"
fastapi = "0.95.0"
opentelemetry-api = "1.36.0"
opentelemetry-sdk = "1.36.0"
opentelemetry-exporter-otlp-proto-grpc = "1.36.0"
opentelemetry-instrumentation-fastapi = "0.57b0"
opentelemetry-instrumentation-grpc = "0.57b0"
opentelemetry-instrumentation-requests = "0.57b0"
prometheus-client = "0.22.1"
huggingface_hub = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
